# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "AccessRequestCreate": {
        "description": "Ask to be put in a group.",
        "properties": {
          "group_id": {
            "format": "uuid",
            "title": "Group Id",
            "type": "string"
          },
          "reason": {
            "description": "Why the access is needed. Required \u2014 an approver with no reason in front of them is rubber-stamping rather than deciding.",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "group_id",
          "reason"
        ],
        "title": "AccessRequestCreate",
        "type": "object"
      },
      "AccessRequestOut": {
        "description": "One request, and what happened to it.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decided_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "decided_by_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By Label"
          },
          "decision_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Note"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "group_id": {
            "format": "uuid",
            "title": "Group Id",
            "type": "string"
          },
          "group_label": {
            "title": "Group Label",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_open": {
            "title": "Is Open",
            "type": "boolean"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "requester_id": {
            "format": "uuid",
            "title": "Requester Id",
            "type": "string"
          },
          "requester_label": {
            "title": "Requester Label",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/RequestState"
          }
        },
        "required": [
          "id",
          "state",
          "requester_id",
          "requester_label",
          "group_id",
          "group_label",
          "reason",
          "decided_by_label",
          "decided_at",
          "decision_note",
          "expires_at",
          "created_at",
          "is_open"
        ],
        "title": "AccessRequestOut",
        "type": "object"
      },
...
      "AccessSummary": {
        "description": "Everything one person has, and where it came from.\n\nThe access review view for a single person. Answers the three questions in\norder: what can they do here, what apps can they get into, and why.",
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "grant_history": {
            "description": "Every role they have ever had, newest first.",
            "items": {
              "$ref": "#/components/schemas/RoleGrantOut"
            },
            "title": "Grant History",
            "type": "array"
          },
          "groups": {
            "description": "Groups they are in.",
            "items": {
              "type": "string"
            },
            "title": "Groups",
            "type": "array"
          },
          "role": {
            "$ref": "#/components/schemas/PlatformRole",
            "description": "What they can do in this console right now."
          },
          "role_expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Expires At"
          },
          "role_granted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Granted At"
          },
          "role_granted_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Who gave them that, if it was granted rather than default.",
            "title": "Role Granted By"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          },
          "user_name": {
            "title": "User Name",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "user_name",
          "display_name",
          "active",
          "role"
        ],
        "title": "AccessSummary",
        "type": "object"
      },
...
      "Decision": {
        "description": "An approver's answer.",
        "properties": {
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Approvals only. When the access should end. 'Until the end of the quarter' is the usual honest answer to an access request.",
            "title": "Expires At"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "What you decided and why. The most useful field here during a review, and the one most likely to be left empty.",
            "title": "Note"
          }
        },
        "title": "Decision",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlatformRole": {
        "description": "What someone can do inside this console.\n\nFour broad roles. Which one someone has is decided by their role grants, not\nby the column on the user \u2014 that column is a cached copy, rebuilt whenever a\ngrant changes. See iam/access/roles.py for why it works that way.\n\nEMPLOYEE is the odd one out: it isn't granted, it's what someone is when\nnothing has been granted to them. So there is never a role grant saying\n\"employee\", and asking for one is refused rather than quietly stored.",
        "enum": [
          "admin",
          "helpdesk",
          "auditor",
          "employee"
        ],
        "title": "PlatformRole",
        "type": "string"
      },
...
      "RequestState": {
        "description": "Where an access request has got to.\n\nEvery state after PENDING is final. A request is a record of somebody asking\nand somebody answering, so reopening one would make \"who approved this\" have\nmore than one answer.",
        "enum": [
          "pending",
          "approved",
          "denied",
          "withdrawn",
          "cancelled"
        ],
        "title": "RequestState",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "SAML 2.0 and SCIM 2.0 identity control plane",
    "title": "IAM Control Plane",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/access-requests": {
      "get": {
        "description": "Open requests, oldest first \u2014 the order they should be worked in.",
        "operationId": "queue_api_access_requests_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AccessRequestOut"
                  },
                  "title": "Response Queue Api Access Requests Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The requests waiting to be decided",
        "tags": [
          "access requests"
        ]
      },
      "post": {
        "description": "Raise a request for yourself.\n\nNo permission required, on purpose. An employee holds no permissions, and a\nrequest system they cannot use is not a request system. Asking grants nothing.\n\nRaises:\n    HTTPException: 400 if the request doesn't make sense, 404 for a group that\n        isn't there.",
        "operationId": "create_request_api_access_requests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessRequestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessRequestOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ask for access to a group",
        "tags": [
          "access requests"
        ]
      }
    },
    "/api/access-requests/group/{group_id}": {
      "get": {
        "description": "The history for a group, decided ones included.\n\nDenied requests are the interesting ones here. \"Three people asked for this and\nwere all refused\" says something about the group that its membership list does\nnot.",
        "operationId": "for_group_api_access_requests_group__group_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Group Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AccessRequestOut"
                  },
                  "title": "Response For Group Api Access Requests Group  Group Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Every request ever raised for one group",
        "tags": [
          "access requests"
        ]
      }
    },
...
    "/api/access-requests/{request_id}": {
      "get": {
        "description": "Read one request.\n\nYours always; anybody else's needs groups:read. \"Who has been asking for the\nfinance system\" is review information, not public.",
        "operationId": "get_request_api_access_requests__request_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessRequestOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One request",
        "tags": [
          "access requests"
        ]
      }
    },
    "/api/access-requests/{request_id}/approve": {
      "post": {
        "description": "Approve, and put them in the group in the same transaction.\n\nRaises:\n    HTTPException: 400 if it can't be approved \u2014 already decided, self-approval,\n        an expiry in the past, or the requester has since left.",
        "operationId": "approve_request_api_access_requests__request_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Decision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessRequestOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve a request and grant the access",
        "tags": [
          "access requests"
        ]
      }
    },
    "/api/access-requests/{request_id}/deny": {
      "post": {
        "description": "Refuse a request, and keep the record of having refused it.",
        "operationId": "deny_request_api_access_requests__request_id__deny_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Decision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessRequestOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Turn a request down",
        "tags": [
          "access requests"
        ]
      }
    },
    "/api/access-requests/{request_id}/withdraw": {
      "post": {
        "description": "Withdraw your own request.\n\nNo permission needed, because it is your own. Somebody else closing it is a\ndenial and goes through the other endpoint \u2014 the two are different answers to\nthe same question and the record should say which happened.",
        "operationId": "withdraw_request_api_access_requests__request_id__withdraw_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessRequestOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Take back a request you raised",
        "tags": [
          "access requests"
        ]
      }
    },
...
    "/api/users/{user_id}/access": {
      "get": {
        "description": "One person's access, gathered in one place.",
        "operationId": "access_summary_api_users__user_id__access_get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Everything this person has, and why",
        "tags": [
          "access"
        ]
      }
    },
...
  }
}